finstat-pro/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.4
numpy==1.24.3
openai==1.3.7
python-dotenv==1.0.0
pydantic==2.5.2
sqlalchemy==2.0.23
alembic==1.13.1
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib==1.7.4
aiofiles==23.2.0
httpx==0.25.2
regex==2023.10.3
nltk==3.8.1
textstat==0.7.3